requirements/backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.16
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
pillow==10.4.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pytz==2024.1
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.1
twisted==24.7.0
txaio==23.1.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
django-prometheus